+++ src/composables/pilots.ts
@@ -115,6 +115,7 @@ export const aircraftSvgColors = (): Record<MapAircraftStatus, string> => {
     return {
         active: getCurrentThemeHexColor('warning700'),
         default: getCurrentThemeHexColor('primary500'),
+        ground: getCurrentThemeHexColor('primary500'),
         green: getCurrentThemeHexColor('success500'),
         hover: getCurrentThemeHexColor('warning600'),
         neutral: getCurrentThemeHexColor('lightgray150'),
@@ -156,7 +157,7 @@ function svgToDataURI(svg: string) {
     return `data:image/svg+xml,${ encoded }`;
 }
 
-export type MapAircraftStatus = 'default' | 'green' | 'active' | 'hover' | 'neutral' | 'arriving' | 'departing' | 'landed';
+export type MapAircraftStatus = 'default' | 'ground' | 'green' | 'active' | 'hover' | 'neutral' | 'arriving' | 'departing' | 'landed';
 
 export async function fetchAircraftIcon(icon: AircraftIcon) {
     const store = useStore();
@@ -199,8 +200,8 @@ export async function loadAircraftIcon({ feature, icon, status, style, rotation,
         image.setRotation(rotation);
     }
     else {
-        if (status === 'default') {
-            const color = store.mapSettings.colors?.[store.getCurrentTheme]?.aircraft?.main;
+        if (status === 'default' || status === 'ground') {
+            const color = store.mapSettings.colors?.[store.getCurrentTheme]?.aircraft?.[status === 'ground' ? 'ground' : 'main'];
             style.setImage(new Icon({
                 src: `/aircraft/${ icon }${ (color && color.color !== 'primary500') ? '-white' : '' }${ store.theme === 'light' ? '-light' : '' }.png?v=${ store.version }`,
                 width: radarIcons[icon].width * (store.mapSettings.aircraftScale ?? 1),